+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-taglib/src/main/java/com/liferay/dynamic/data/mapping/form/taglib/internal/servlet/taglib/util/DDMFormTaglibUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.form.taglib.servlet.taglib.util;
+package com.liferay.dynamic.data.mapping.form.taglib.internal.servlet.taglib.util;
 
 import com.liferay.dynamic.data.mapping.form.builder.context.DDMFormBuilderContextFactory;
 import com.liferay.dynamic.data.mapping.form.builder.context.DDMFormBuilderContextRequest;
