+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-taglib/src/test/java/com/liferay/dynamic/data/mapping/form/taglib/internal/servlet/taglib/util/DDMFormTaglibUtilTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.form.taglib.servlet.taglib.util;
+package com.liferay.dynamic.data.mapping.form.taglib.internal.servlet.taglib.util;
 
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
