+++ modules/apps/flags/flags-taglib/src/main/java/com/liferay/flags/taglib/internal/servlet/taglib/util/FlagsTagUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.flags.taglib.servlet.taglib.util;
+package com.liferay.flags.taglib.internal.servlet.taglib.util;
 
 import com.liferay.flags.configuration.FlagsGroupServiceConfiguration;
 import com.liferay.portal.kernel.exception.PortalException;
