+++ modules/apps/flags/flags-taglib/src/main/java/com/liferay/flags/taglib/servlet/taglib/react/FlagsTag.java
@@ -15,7 +15,7 @@
 package com.liferay.flags.taglib.servlet.taglib.react;
 
 import com.liferay.flags.taglib.internal.servlet.ServletContextUtil;
-import com.liferay.flags.taglib.servlet.taglib.util.FlagsTagUtil;
+import com.liferay.flags.taglib.internal.servlet.taglib.util.FlagsTagUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
